Refactor evaluation logic into EvaluationOrchestrator
