+++ framework/evaluation.py
@@ -1,34 +1,3 @@
-from __future__ import annotations
-
-import asyncio
-import importlib.util
-import logging
-import sys
-from pathlib import Path
-from types import ModuleType
-from typing import Any, Dict, List, Optional, Tuple
-
-from agents import trace
-from pydantic import BaseModel, Field
-from rich.console import Console
-from rich.table import Table
-from rich.text import Text
-
-from config import get_external_client
-from framework.declarative_agents import (
-    Agent,
-    AgentLoader,
-    AgentSpecification,
-    ModelSettings,
-    OpenAIChatCompletionsModel,
-    Runner,
-)
-from framework.types import AgentType, StructuredOutputAgentDefinition
-from framework.utils import ThinkTagFilter, extract_text_delta_from_event
-
-
-
-
 """
 Evaluation utilities for the declarative agent framework.
 
@@ -48,41 +17,18 @@ The framework discovers these variables reflectively and evaluates the agent
 against each case using an AI-judge pattern.
 """
 
-console = Console()
+from __future__ import annotations
 
-logger = logging.getLogger(__name__)
-_EVAL_LOGGING_CONFIGURED: bool = False
+import asyncio
+import importlib.util
+from pathlib import Path
+from types import ModuleType
+from typing import Any, Dict, List, Tuple
 
-# --- Models -----------------------------------------------------------------
+from agents import trace
 
-
-class EvaluationResult(BaseModel):
-    """Schema returned by the judge formatter."""
-
-    passed: bool = Field(description="True if the agent's output meets all criteria, False otherwise.")
-    reasoning: str = Field(description="Explanation referencing the specific criteria that were met or failed.")
-
-
-# --- Helper functions -------------------------------------------------------
-
-from config import BIG_MODEL, Model, big_model_settings
-
-_QWEN_MODEL_NAME: Model = BIG_MODEL
-_QWEN_MODEL_SETTINGS: ModelSettings = big_model_settings
-
-_client = get_external_client()
-if _client is None:
-    logger.warning("LM Studio client not detected – evaluations that rely on external LLMs will fail")
-
-# Lazily-constructed global judge agents so we don't spawn them repeatedly
-_judge_agent: Optional[Agent] = None
-_judge_formatter_agent: Optional[Agent] = None
-
-
-# ---------------------------------------------------------------------------
-# Evaluation helpers
-# ---------------------------------------------------------------------------
-
+from framework.declarative_agents import AgentLoader, AgentSpecification
+from framework.evaluation_orchestrator import EvaluationOrchestrator
 
 async def evaluate_agent_against_suite(
     agent_spec: AgentSpecification,
@@ -91,243 +37,10 @@ async def evaluate_agent_against_suite(
 ) -> None:
     """Run a test suite against an agent specification and print a summary.
 
-    If the YAML defines `type: "structured_output"` we invoke the full two-step
-    `StructuredOutputAgent` chain so the formatter model guarantees valid JSON.
-    Otherwise we fall back to the original direct-agent execution path.
-
-    The judge now evaluates the agent output independently against each criterion.
+    This is a convenience wrapper around EvaluationOrchestrator for backward compatibility.
     """
-
-    # Cached judge agents are not used for per-criterion evaluation to avoid any hidden state carryover.
-
-    use_structured_output = isinstance(agent_spec.definition, StructuredOutputAgentDefinition)
-
-    # Non-structured agents are materialised once and reused across cases – this
-    # mirrors the original implementation.
-    direct_agent: Optional[Agent] = None
-    if not use_structured_output:
-        direct_agent = agent_spec.create_agent()
-
-    target_agent_name = (
-        agent_spec.definition.name
-        if use_structured_output
-        else direct_agent.name if direct_agent is not None else "unknown"
-    )
-
-    # Create logs directory once at the start and configure logging sinks
-    logs_dir = Path("evaluation_logs") / target_agent_name
-    _configure_evaluation_logging(logs_dir)
-
-    logger.info("--- Starting Evaluation for: %s ---", target_agent_name)
-    pass_count = 0
-    total_count = len(test_suite)
-
-    # Data structures to track UI state and logs
-    case_rows: List[Dict[str, str]] = []
-
-    for idx, test_case in enumerate(test_suite, start=1):
-        test_id: str = test_case.get("id", f"case_{idx}")
-        agent_input: str = test_case.get("prompt", "")
-
-        parsed_output = None
-
-        if use_structured_output:
-            # ----------------------------------------------------------
-            # Use the full StructuredOutputAgent chain so that the
-            # formatter model produces clean, parseable JSON.
-            # ----------------------------------------------------------
-            try:
-                result_model = await agent_spec.structured_output_agent.run(
-                    agent_input,
-                )
-                # Convert the Pydantic model into a regular dict so that
-                # downstream code (judge prompt) can embed it easily.
-                parsed_output = result_model.model_dump()
-            except Exception as e:
-                logger.exception("Error running StructuredOutputAgent: %s", e)
-                parsed_output = None
-        else:
-            # ----------------------------------------------------------
-            # Legacy path – run the single agent directly with streaming.
-            # This mirrors the live streaming used by non-eval runs so
-            # think tokens are surfaced to stdout and captured in logs.
-            # ----------------------------------------------------------
-            assert direct_agent is not None  # Mypy guard
-
-            # Build kwargs for Runner.run_streamed to honour max_iterations if specified
-            max_turns = 10
-            if agent_spec.definition.max_iterations is not None:
-                max_turns = agent_spec.definition.max_iterations
-
-            streamed = Runner.run_streamed(
-                starting_agent=direct_agent,
-                input=agent_input,
-                max_turns=max_turns,
-            )
-
-            think_filter = ThinkTagFilter()
-            should_print_think = bool(agent_spec.definition.print_think_tokens)
-
-            raw_chunks: list[str] = []
-            visible_chunks: list[str] = []
-
-            async for event in streamed.stream_events():
-                text_delta = extract_text_delta_from_event(event)
-                if text_delta is None:
-                    continue
-
-                raw_chunks.append(text_delta)
-                if should_print_think:
-                    print(text_delta, end="", flush=True)
-                    visible_chunks.append(text_delta)
-                else:
-                    chunk = think_filter.filter_token(text_delta)
-                    if chunk:
-                        print(chunk, end="", flush=True)
-                        visible_chunks.append(chunk)
-
-            # Ensure newline after stream
-            print()
-
-            # Persist both raw and visible streams into the evaluation log
-            raw_stream = "".join(raw_chunks)
-            visible_stream = "".join(visible_chunks)
-            if raw_stream:
-                logger.info("Raw stream output (including <think>):\n%s", raw_stream)
-            if visible_stream and visible_stream != raw_stream:
-                logger.info("Visible stream output (after think filtering):\n%s", visible_stream)
-
-            # Use the schema-derived model decided at spec load time
-            agent_stream = streamed
-            if agent_spec.output_model is not None:
-                parsed_output = agent_stream.final_output_as(agent_spec.output_model)
-
-        if parsed_output is None:
-            case_rows.append({"id": test_id, "result": "FAIL", "criteria": "0/0"})
-            logger.debug("No parseable output produced by the agent; skipping judge phase.")
-            continue
-
-        # Judge per criterion independently
-        criteria_list = [str(c).strip() for c in (evaluation_criteria or []) if str(c).strip()]
-        if not criteria_list:
-            case_rows.append({"id": test_id, "result": "FAIL", "criteria": "0/0"})
-            logger.warning("[FAIL] %s – empty criteria list provided; cannot judge.", test_id)
-            continue
-
-        per_criterion_results: List[EvaluationResult] = []
-
-        for crit_index, single_criterion in enumerate(criteria_list, start=1):
-            judge_agent, judge_formatter_agent = _build_fresh_judge_agents()
-            judge_prompt = f"""The current date is June 2025.
-
-Please evaluate the following agent output for logical correctness.
-The output has already been validated as parsable. You only need to check the content against the single evaluation criterion below.
-
-Evaluation Criterion:\n```\n{single_criterion}\n```
-
-**Original Prompt:**\n```\n{agent_input}\n```\n\n**Agent's Parsed Output (as JSON):**\n```json\n{parsed_output}\n```\n\n**Test Case Details (for context):**\n```json\n{test_case}\n```"""
-
-            # Stream judge output to stdout and log file
-            judge_streamed = Runner.run_streamed(
-                starting_agent=judge_agent,
-                input=judge_prompt,
-            )
-
-            judge_raw_chunks: list[str] = []
-            async for event in judge_streamed.stream_events():
-                text_delta = extract_text_delta_from_event(event)
-                if text_delta is None:
-                    continue
-                print(text_delta, end="", flush=True)
-                judge_raw_chunks.append(text_delta)
-            print()
-            if judge_raw_chunks:
-                logger.info(
-                    "Judge raw stream output for criterion %s (token-by-token):\n%s",
-                    crit_index,
-                    "".join(judge_raw_chunks),
-                )
-
-            judge_output_text = str(judge_streamed.final_output)
-
-            # Stream formatter output as well
-            formatter_streamed = Runner.run_streamed(
-                starting_agent=judge_formatter_agent,
-                input=judge_output_text,
-            )
-
-            formatter_raw_chunks: list[str] = []
-            async for event in formatter_streamed.stream_events():
-                text_delta = extract_text_delta_from_event(event)
-                if text_delta is None:
-                    continue
-                print(text_delta, end="", flush=True)
-                formatter_raw_chunks.append(text_delta)
-            print()
-            if formatter_raw_chunks:
-                logger.info(
-                    "Judge formatter raw stream output for criterion %s (token-by-token):\n%s",
-                    crit_index,
-                    "".join(formatter_raw_chunks),
-                )
-
-            eval_result = formatter_streamed.final_output_as(EvaluationResult)
-
-            if eval_result is None:
-                logger.warning(
-                    "[FAIL] %s – judge formatter could not parse result for criterion %s; raw: %s",
-                    test_id,
-                    crit_index,
-                    formatter_streamed.final_output,
-                )
-                per_criterion_results.append(
-                    EvaluationResult(passed=False, reasoning="Formatter could not parse result")
-                )
-            else:
-                per_criterion_results.append(eval_result)
-                level = logging.INFO if eval_result.passed else logging.WARNING
-                logger.log(level, "Criterion %s passed: %s", crit_index, eval_result.passed)
-                logger.info("Criterion %s reasoning: %s", crit_index, eval_result.reasoning)
-
-        # Aggregate: overall pass only if all criteria passed
-        overall_pass = all(r.passed for r in per_criterion_results) if per_criterion_results else False
-        passed_count = sum(1 for r in per_criterion_results if r.passed)
-        total_criteria = len(per_criterion_results)
-        failed_indices = [str(i + 1) for i, r in enumerate(per_criterion_results) if not r.passed]
-        criteria_summary = f"{passed_count}/{total_criteria}" + (
-            f" (failed: {', '.join(failed_indices)})" if failed_indices else ""
-        )
-
-        if overall_pass:
-            pass_count += 1
-            case_rows.append({"id": test_id, "result": "PASS", "criteria": criteria_summary})
-            logger.info("[PASS] Test ID: %s (all %s criteria passed)", test_id, len(per_criterion_results))
-        else:
-            case_rows.append({"id": test_id, "result": "FAIL", "criteria": criteria_summary})
-            logger.warning("[FAIL] Test ID: %s (failed criteria: %s)", test_id, ", ".join(failed_indices) or "n/a")
-
-    # Summary
-    table = Table(show_header=True, header_style="bold magenta")
-    table.add_column("Test ID", width=30, no_wrap=True)
-    table.add_column("Result", justify="right", width=10)
-    table.add_column("Criteria", justify="left", width=30)
-    for row in case_rows:
-        result_text = Text(row["result"])
-        if row["result"] == "PASS":
-            result_text.stylize("bold green")
-        else:
-            result_text.stylize("bold red")
-        table.add_row(row["id"], result_text, row.get("criteria", ""))
-
-    console.print(table)
-
-    pass_rate = (pass_count / total_count) * 100 if total_count > 0 else 0
-    logger.info(
-        "--- Evaluation Summary — Passed: %s / %s (%.2f%%) ---",
-        pass_count,
-        total_count,
-        pass_rate,
-    )
+    orchestrator = EvaluationOrchestrator(agent_spec)
+    await orchestrator.evaluate_against_suite(test_suite, evaluation_criteria)
 
 
 # --- Discovery helpers ------------------------------------------------------
@@ -395,135 +108,6 @@ def _extract_suite_and_criteria(module: ModuleType) -> Tuple[List[Dict[str, Any]
     return test_suite, criteria
 
 
-# --- Helper functions -------------------------------------------------------
-
-
-def _configure_evaluation_logging(logs_dir: Path) -> None:
-    """Ensure evaluation logs go to both stdout and a rotating file.
-
-    The configuration is idempotent; repeated calls will be no-ops.
-    """
-    global _EVAL_LOGGING_CONFIGURED
-    if _EVAL_LOGGING_CONFIGURED:
-        return
-
-    logs_dir.mkdir(parents=True, exist_ok=True)
-
-    logger.setLevel(logging.INFO)
-    logger.propagate = False
-
-    formatter = logging.Formatter(
-        fmt="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
-        datefmt="%Y-%m-%d %H:%M:%S",
-    )
-
-    # File handler
-    file_handler = logging.FileHandler(logs_dir / "evaluation.log", mode="a", encoding="utf-8")
-    file_handler.setLevel(logging.INFO)
-    file_handler.setFormatter(formatter)
-
-    # Stream handler to stdout
-    stream_handler = logging.StreamHandler(sys.stdout)
-    stream_handler.setLevel(logging.INFO)
-    stream_handler.setFormatter(formatter)
-
-    logger.addHandler(file_handler)
-    logger.addHandler(stream_handler)
-
-    _EVAL_LOGGING_CONFIGURED = True
-
-
-def _get_judge_agents() -> Tuple[Agent, Agent]:
-    """Create (or return cached) judge + formatter agents."""
-
-    global _judge_agent, _judge_formatter_agent
-
-    if _judge_agent and _judge_formatter_agent:
-        return _judge_agent, _judge_formatter_agent
-
-    judge_instructions = """
-You are a meticulous and impartial AI evaluator. Your task is to assess the output of another AI agent based on a given set of criteria.
-
-You will be provided with:
-1.  The `agent_output` that the agent produced.
-2.  The `evaluation_criteria` that you must strictly follow to determine if the test passes or fails.
-3.  The `test_case` details for context.
-
-CRITICAL RULE: Your decision about whether the test passes or fails MUST be based *exclusively* on the `evaluation_criteria`. The `test_case` (including the original `prompt`) is provided for context only. Do NOT infer the agent's expected behavior from the `prompt`; use only the `evaluation_criteria`.
-
-Your process is to:
-1.  **Understand the Goal**: Read the `evaluation_criteria` carefully to understand what constitutes a "pass".
-2.  **Analyze the Output**: Scrutinize the `agent_output` and compare it against the rules in the `evaluation_criteria`.
-3.  **Make a Decision**: Based *only* on the `evaluation_criteria`, decide if the `agent_output` passes the test.
-4.  **Formulate Reasoning**: In the `reasoning` field, explain *why* the test passed or failed by referencing the specific rule(s) from the `evaluation_criteria` that were met or not met.
-5.  **Set Final Verdict**: Set the `passed` field to `true` if the test passed, and `false` if it failed.
-
-You may use <think> tags to reason about the query before producing your final output. Your final output must be a single, valid JSON object with 'passed' (boolean) and 'reasoning' (string) fields.
-"""
-
-    _judge_agent = Agent(
-        name="EvaluationJudgeAgent",
-        instructions=judge_instructions,
-        model=OpenAIChatCompletionsModel(model=_QWEN_MODEL_NAME, openai_client=_client),
-        model_settings=_QWEN_MODEL_SETTINGS,
-    )
-
-    _judge_formatter_agent = Agent(
-        name="EvaluationJudgeFormatterAgent",
-        instructions=(
-            "You are a formatter agent. Your task is to convert the user's input text into a valid JSON "
-            "object conforming to the EvaluationResult schema. The schema requires a 'passed' field (boolean) "
-            "and a 'reasoning' field (string). Your output MUST be ONLY the JSON object."
-        ),
-        model=OpenAIChatCompletionsModel(model=_QWEN_MODEL_NAME, openai_client=_client),
-        model_settings=_QWEN_MODEL_SETTINGS,
-        output_type=EvaluationResult,
-    )
-
-    return _judge_agent, _judge_formatter_agent
-
-
-def _build_fresh_judge_agents() -> Tuple[Agent, Agent]:
-    """Construct fresh judge and formatter agents to ensure independence per criterion."""
-    judge_instructions = """
-You are a meticulous and impartial AI evaluator. Your task is to assess the output of another AI agent based on a given set of criteria.
-
-You will be provided with:
-1.  The `agent_output` that the agent produced.
-2.  The `evaluation_criteria` that you must strictly follow to determine if the test passes or fails.
-3.  The `test_case` details for context.
-
-CRITICAL RULE: Your decision about whether the test passes or fails MUST be based *exclusively* on the `evaluation_criteria`. The `test_case` (including the original `prompt`) is provided for context only. Do NOT infer the agent's expected behavior from the `prompt`; use only the `evaluation_criteria`.
-
-Your process is to:
-1.  **Understand the Goal**: Read the `evaluation_criteria` carefully to understand what constitutes a "pass".
-2.  **Analyze the Output**: Scrutinize the `agent_output` and compare it against the rules in the `evaluation_criteria`.
-3.  **Make a Decision**: Based *only* on the `evaluation_criteria`, decide if the `agent_output` passes the test.
-4.  **Formulate Reasoning**: In the `reasoning` field, explain *why* the test passed or failed by referencing the specific rule(s) from the `evaluation_criteria` that were met or not met.
-5.  **Set Final Verdict**: Set the `passed` field to `true` if the test passed, and `false` if it failed.
-
-You may use <think> tags to reason about the query before producing your final output. Your final output must be a single, valid JSON object with 'passed' (boolean) and 'reasoning' (string) fields.
-"""
-    judge = Agent(
-        name="EvaluationJudgeAgent",
-        instructions=judge_instructions,
-        model=OpenAIChatCompletionsModel(model=_QWEN_MODEL_NAME, openai_client=_client),
-        model_settings=_QWEN_MODEL_SETTINGS,
-    )
-
-    formatter = Agent(
-        name="EvaluationJudgeFormatterAgent",
-        instructions=(
-            "You are a formatter agent. Your task is to convert the user's input text into a valid JSON "
-            "object conforming to the EvaluationResult schema. The schema requires a 'passed' field (boolean) "
-            "and a 'reasoning' field (string). Your output MUST be ONLY the JSON object."
-        ),
-        model=OpenAIChatCompletionsModel(model=_QWEN_MODEL_NAME, openai_client=_client),
-        model_settings=_QWEN_MODEL_SETTINGS,
-        output_type=EvaluationResult,
-    )
-
-    return judge, formatter
 
 
 # --- Public entrypoint ------------------------------------------------------
